fix a bug: delete all blogtags when update a blog without tags

+++ growing-pains-service/src/main/java/com/growing/pains/service/blog/BlogService.java
@@ -75,7 +75,10 @@ public class BlogService {
     @Transactional
     public void updateBlogTag(List<Integer> tagIds, int blogContentId) {
         List<BlogContentTagEntity> entities = queryTagByContentId(blogContentId);
-        blogContentTagDao.delete(entities);
+        if (!CollectionUtils.isEmpty(entities)) {
+            blogContentTagDao.delete(entities);
+        }
+
         if (CollectionUtils.isEmpty(tagIds)) {
             return;
         }
